+++ CHANGELOG.md
@@ -6,8 +6,9 @@ None
 
 ## [WIP] 0.3.0 / 2019-January-??
 
-- [Todo] Add inf-clojure (Inferior Clojure mode)
-- [Todo] Disable CIDER by default
+- Add inf-clojure (Inferior Clojure mode)
+- Enable inf-clojure by default
+- Disable CIDER by default
 
 
 ## 0.2.1 / 2017-December-31